Refactored Player Movement
Movement now in separate character controller script.  Started implementing a ground check for jumping
