+++ EscapeForest/Assets/CharacterController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 881c9e4b599ee05469acf54bb527e632
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
