+++ EscapeForest/ProjectSettings/TagManager.asset
@@ -44,6 +44,12 @@ TagManager:
   - name: Default
     uniqueID: 0
     locked: 0
+  - name: Ground
+    uniqueID: 3279804563
+    locked: 0
+  - name: Player
+    uniqueID: 1820408553
+    locked: 0
   - name: Foreground
     uniqueID: 3478511521
     locked: 0
